little rearrangement
