+++ root/app/snippies/db.py
@@ -2,6 +2,7 @@ import sqlite3
 from datetime import datetime
 from snippies import app_log
 
+
 def init_db(database_path):
     app_log.debug("Initializing backup schedules database")
     with sqlite3.connect(database_path) as db:
@@ -14,6 +15,7 @@ def init_db(database_path):
             );
         """)
 
+
 def get_records():
     app_log.debug("Getting backup records")
     with sqlite3.connect('schedules.sqlite') as database:
@@ -25,15 +27,17 @@ def get_records():
         else:
             return rows
 
+
 def add_record(username, date, action):
     app_log.debug("Formatting datetime object to string")
     date = datetime.strftime(date, "%Y-%m-%dT%H:%M:%S.000Z")
     app_log.debug(f"Adding record ({username}, {date}, {action}) into database")
     with sqlite3.connect('schedules.sqlite') as database:
         cur = database.execute("INSERT INTO schedules (username, date, action) VALUES (?, ?, ?)",
-            (username, date, action))
+                               (username, date, action))
         return cur.lastrowid
-    
+
+
 def delete_record(row_id):
     app_log.debug(f"Attempting to delete record {row_id}")
     with sqlite3.connect('schedules.sqlite') as database:
