+++ root/app/snippies/scheduler_config.py
@@ -2,7 +2,7 @@ from apscheduler.jobstores.sqlalchemy import SQLAlchemyJobStore
 from apscheduler.executors.pool import ThreadPoolExecutor, ProcessPoolExecutor
 
 jobstores = {
-    'default': SQLAlchemyJobStore(url='sqlite:///jobs.sqlite')
+    'default': SQLAlchemyJobStore(url='sqlite:///config/jobs.sqlite')
 }
 executors = {
     'default': ThreadPoolExecutor(20),
