+++ root/etc/s6-overlay/s6-rc.d/init-config-adoverseas/run
@@ -1,0 +1,5 @@
+#!/usr/bin/with-contenv bash
+
+# permissions
+chown -R abc:abc \
+	/config
