+++ root/etc/s6-overlay/s6-rc.d/svc-adoverseas/run
@@ -1,0 +1,7 @@
+#!/usr/bin/with-contenv bash
+
+cd /app || :
+
+exec \
+    s6-setuidgid abc \
+        flask --app ADOverseas run --host 0.0.0.0 --port 3500
