+++ root/etc/s6-overlay/s6-rc.d/svc-adoverseas/type
@@ -1,0 +1,1 @@
+longrun
